Move scanner implementations to `src/scanner`
