--- src/arch/arm/Frequency.hpp
@@ -1,117 +1,0 @@
-#pragma once
-
-#include <array>
-#include <cstddef>
-#include <cstdint>
-#include <utility>
-
-namespace hat::detail::aarch64 {
-
-    static constexpr auto p(std::uint8_t a, std::uint8_t b) {
-        return std::pair{std::byte{a}, std::byte{b}};
-    }
-
-    // Top 512 byte pair occurrences on 1 byte alignment, sorted. Sourced from all PE (.exe and .dll) files included in
-    // the "clang+llvm-22.1.8-aarch64-pc-windows-msvc" archive. This list accounts for ~50.9% of all byte pairs.
-    static constexpr inline auto pairs_x1 = std::to_array<std::pair<std::byte, std::byte>>({
-        p(0x00, 0x00), p(0x00, 0x01), p(0x00, 0x08), p(0x00, 0x11), p(0x00, 0x12), p(0x00, 0x14), p(0x00, 0x20), p(0x00, 0x2A),
-        p(0x00, 0x34), p(0x00, 0x35), p(0x00, 0x36), p(0x00, 0x37), p(0x00, 0x39), p(0x00, 0x40), p(0x00, 0x51), p(0x00, 0x54),
-        p(0x00, 0x71), p(0x00, 0x79), p(0x00, 0x80), p(0x00, 0x90), p(0x00, 0x91), p(0x00, 0x94), p(0x00, 0xA9), p(0x00, 0xAA),
-        p(0x00, 0xB0), p(0x00, 0xB4), p(0x00, 0xB5), p(0x00, 0xB9), p(0x00, 0xD0), p(0x00, 0xD1), p(0x00, 0xF0), p(0x00, 0xF1),
-        p(0x00, 0xF9), p(0x00, 0xFD), p(0x01, 0x00), p(0x01, 0x01), p(0x01, 0x05), p(0x01, 0x08), p(0x01, 0x09), p(0x01, 0x0A),
-        p(0x01, 0x0B), p(0x01, 0x0C), p(0x01, 0x15), p(0x01, 0x3F), p(0x01, 0x40), p(0x01, 0x54), p(0x01, 0x71), p(0x01, 0x80),
-        p(0x01, 0x90), p(0x01, 0x91), p(0x01, 0xA9), p(0x01, 0xAA), p(0x01, 0xB0), p(0x01, 0xB9), p(0x01, 0xD0), p(0x01, 0xD1),
-        p(0x01, 0xF0), p(0x01, 0xF9), p(0x02, 0x00), p(0x02, 0x01), p(0x02, 0x05), p(0x02, 0x08), p(0x02, 0x40), p(0x02, 0x80),
-        p(0x02, 0x91), p(0x02, 0xA9), p(0x02, 0xAA), p(0x02, 0xF9), p(0x03, 0x00), p(0x03, 0x01), p(0x03, 0x02), p(0x03, 0x03),
-        p(0x03, 0x04), p(0x03, 0x05), p(0x03, 0x08), p(0x03, 0x09), p(0x03, 0x13), p(0x03, 0x14), p(0x03, 0x15), p(0x03, 0x16),
-        p(0x03, 0x17), p(0x03, 0x18), p(0x03, 0x19), p(0x03, 0x1A), p(0x03, 0x1B), p(0x03, 0x1C), p(0x03, 0x1F), p(0x03, 0x40),
-        p(0x03, 0x5F), p(0x03, 0x80), p(0x03, 0x91), p(0x03, 0xA9), p(0x03, 0xAA), p(0x04, 0x00), p(0x04, 0x40), p(0x04, 0x80),
-        p(0x04, 0x91), p(0x04, 0xA9), p(0x05, 0x00), p(0x05, 0x01), p(0x05, 0x40), p(0x05, 0x91), p(0x06, 0x00), p(0x06, 0x40),
-        p(0x06, 0x80), p(0x06, 0x91), p(0x07, 0x00), p(0x07, 0x40), p(0x07, 0x80), p(0x07, 0x91), p(0x08, 0x00), p(0x08, 0x01),
-        p(0x08, 0x05), p(0x08, 0x09), p(0x08, 0x10), p(0x08, 0x11), p(0x08, 0x21), p(0x08, 0x40), p(0x08, 0x41), p(0x08, 0x61),
-        p(0x08, 0x81), p(0x08, 0x8B), p(0x08, 0x91), p(0x08, 0xA1), p(0x08, 0xAA), p(0x08, 0xC1), p(0x08, 0xCB), p(0x08, 0xEB),
-        p(0x09, 0x00), p(0x09, 0x01), p(0x09, 0x05), p(0x09, 0x40), p(0x09, 0x6B), p(0x09, 0x8B), p(0x09, 0x91), p(0x09, 0xEB),
-        p(0x0A, 0x00), p(0x0A, 0x01), p(0x0A, 0x40), p(0x0A, 0xEB), p(0x0B, 0x00), p(0x0B, 0x40), p(0x0C, 0x00), p(0x0D, 0x00),
-        p(0x0D, 0x40), p(0x0E, 0x00), p(0x0E, 0x40), p(0x0F, 0x00), p(0x0F, 0x40), p(0x10, 0x00), p(0x10, 0x40), p(0x11, 0x00),
-        p(0x11, 0x40), p(0x12, 0x00), p(0x12, 0x40), p(0x13, 0x00), p(0x13, 0x40), p(0x13, 0xAA), p(0x14, 0x00), p(0x14, 0xAA),
-        p(0x14, 0xE0), p(0x14, 0xE8), p(0x14, 0xEB), p(0x15, 0x00), p(0x15, 0xAA), p(0x15, 0xEB), p(0x16, 0x40), p(0x16, 0xAA),
-        p(0x17, 0x00), p(0x17, 0x08), p(0x17, 0x40), p(0x17, 0x9F), p(0x17, 0xAA), p(0x17, 0xE0), p(0x17, 0xE8), p(0x18, 0xAA),
-        p(0x19, 0x00), p(0x19, 0xAA), p(0x1A, 0x00), p(0x1A, 0x40), p(0x1A, 0xAA), p(0x1B, 0x00), p(0x1B, 0x40), p(0x1B, 0xAA),
-        p(0x1C, 0xAA), p(0x1D, 0x00), p(0x1F, 0x00), p(0x1F, 0x01), p(0x1F, 0x05), p(0x1F, 0x09), p(0x1F, 0x20), p(0x1F, 0x2A),
-        p(0x1F, 0x40), p(0x1F, 0x41), p(0x1F, 0x7C), p(0x1F, 0xAA), p(0x20, 0x00), p(0x20, 0x01), p(0x20, 0x80), p(0x20, 0xD4),
-        p(0x21, 0x00), p(0x21, 0x40), p(0x22, 0x00), p(0x22, 0x40), p(0x23, 0x00), p(0x23, 0x01), p(0x23, 0x40), p(0x27, 0x00),
-        p(0x27, 0x40), p(0x28, 0x00), p(0x28, 0x01), p(0x29, 0x00), p(0x29, 0x01), p(0x29, 0x05), p(0x2A, 0x00), p(0x2A, 0xE0),
-        p(0x2B, 0x00), p(0x2B, 0x40), p(0x2F, 0x00), p(0x2F, 0x40), p(0x33, 0x00), p(0x33, 0x40), p(0x39, 0x08), p(0x39, 0x1F),
-        p(0x39, 0xE8), p(0x3D, 0x00), p(0x3D, 0xE0), p(0x3F, 0x00), p(0x3F, 0x01), p(0x3F, 0x05), p(0x3F, 0xD6), p(0x40, 0x00),
-        p(0x40, 0x01), p(0x40, 0x39), p(0x40, 0x79), p(0x40, 0x91), p(0x40, 0x92), p(0x40, 0xA9), p(0x40, 0xB1), p(0x40, 0xB9),
-        p(0x40, 0xF2), p(0x40, 0xF8), p(0x40, 0xF9), p(0x40, 0xFD), p(0x41, 0x00), p(0x41, 0x39), p(0x41, 0xA9), p(0x41, 0xB9),
-        p(0x41, 0xF8), p(0x41, 0xF9), p(0x42, 0x00), p(0x42, 0xA9), p(0x42, 0xB9), p(0x42, 0xF9), p(0x43, 0x00), p(0x43, 0x01),
-        p(0x43, 0x40), p(0x43, 0xA9), p(0x43, 0xF9), p(0x44, 0xA9), p(0x44, 0xF9), p(0x45, 0xF9), p(0x4A, 0x01), p(0x51, 0x1F),
-        p(0x52, 0x08), p(0x52, 0x09), p(0x52, 0x1F), p(0x52, 0xE0), p(0x52, 0xE3), p(0x52, 0xE8), p(0x53, 0xC2), p(0x54, 0x08),
-        p(0x54, 0x09), p(0x54, 0x1F), p(0x54, 0x28), p(0x54, 0x29), p(0x54, 0x3F), p(0x54, 0x68), p(0x54, 0x88), p(0x54, 0xE0),
-        p(0x54, 0xE8), p(0x54, 0xE9), p(0x54, 0xFE), p(0x5B, 0x01), p(0x5B, 0x41), p(0x5F, 0x01), p(0x5F, 0xD6), p(0x5F, 0xF8),
-        p(0x60, 0x00), p(0x60, 0x02), p(0x61, 0x00), p(0x62, 0x00), p(0x63, 0x00), p(0x68, 0x00), p(0x68, 0x02), p(0x68, 0x0A),
-        p(0x69, 0x02), p(0x6B, 0x01), p(0x71, 0x61), p(0x7C, 0x92), p(0x7C, 0xD3), p(0x7D, 0x92), p(0x7D, 0xD3), p(0x7F, 0x01),
-        p(0x7F, 0x02), p(0x80, 0x00), p(0x80, 0x3D), p(0x80, 0x52), p(0x81, 0x00), p(0x82, 0x00), p(0x83, 0x00), p(0x83, 0x01),
-        p(0x83, 0x02), p(0x88, 0x00), p(0x88, 0x02), p(0x88, 0x9A), p(0x89, 0x9A), p(0x90, 0x08), p(0x91, 0x00), p(0x91, 0x01),
-        p(0x91, 0x02), p(0x91, 0x03), p(0x91, 0x08), p(0x91, 0x1F), p(0x91, 0x22), p(0x91, 0xC0), p(0x91, 0xE0), p(0x91, 0xE1),
-        p(0x91, 0xE2), p(0x91, 0xE3), p(0x91, 0xE4), p(0x91, 0xE8), p(0x91, 0xE9), p(0x91, 0xFF), p(0x94, 0x20), p(0x94, 0x68),
-        p(0x94, 0xE0), p(0x94, 0xE1), p(0x94, 0xE8), p(0x97, 0x20), p(0x97, 0x68), p(0x97, 0xE0), p(0x97, 0xE1), p(0x97, 0xE8),
-        p(0x9F, 0x02), p(0x9F, 0x1A), p(0xA0, 0x00), p(0xA1, 0x00), p(0xA3, 0x00), p(0xA8, 0x02), p(0xA8, 0xC0), p(0xA9, 0x08),
-        p(0xA9, 0x1F), p(0xA9, 0xE0), p(0xA9, 0xE8), p(0xA9, 0xF3), p(0xA9, 0xF5), p(0xA9, 0xF7), p(0xA9, 0xF9), p(0xA9, 0xFE),
-        p(0xA9, 0xFF), p(0xAA, 0x00), p(0xAA, 0x01), p(0xAA, 0x02), p(0xAA, 0x03), p(0xAA, 0x08), p(0xAA, 0x1F), p(0xAA, 0xE0),
-        p(0xAA, 0xE1), p(0xAA, 0xE2), p(0xAA, 0xE3), p(0xAA, 0xE4), p(0xAA, 0xE5), p(0xAA, 0xE8), p(0xAA, 0xF3), p(0xAA, 0xF4),
-        p(0xAA, 0xF5), p(0xAA, 0xF6), p(0xAA, 0xF7), p(0xAA, 0xFE), p(0xAA, 0xFF), p(0xB0, 0x08), p(0xB4, 0x08), p(0xB4, 0xE0),
-        p(0xB9, 0x08), p(0xB9, 0x09), p(0xB9, 0x1F), p(0xB9, 0x29), p(0xB9, 0x3F), p(0xB9, 0x68), p(0xB9, 0xE0), p(0xB9, 0xE1),
-        p(0xB9, 0xE8), p(0xB9, 0xE9), p(0xB9, 0xFF), p(0xBF, 0x02), p(0xC0, 0x03), p(0xC0, 0x3D), p(0xC1, 0x00), p(0xC2, 0xA8),
-        p(0xC3, 0x00), p(0xC3, 0x01), p(0xC8, 0x02), p(0xD0, 0x08), p(0xD1, 0xF3), p(0xD4, 0xFF), p(0xD6, 0xE0), p(0xDF, 0x02),
-        p(0xE0, 0x03), p(0xE0, 0x07), p(0xE0, 0x0F), p(0xE0, 0x23), p(0xE0, 0x43), p(0xE0, 0x63), p(0xE0, 0x83), p(0xE0, 0xC3),
-        p(0xE1, 0x03), p(0xE1, 0x23), p(0xE1, 0x43), p(0xE1, 0x63), p(0xE1, 0x83), p(0xE1, 0xC3), p(0xE2, 0x03), p(0xE3, 0x00),
-        p(0xE3, 0x03), p(0xE4, 0x00), p(0xE4, 0x03), p(0xE5, 0x03), p(0xE6, 0x03), p(0xE7, 0x03), p(0xE8, 0x03), p(0xE8, 0x07),
-        p(0xE8, 0x0B), p(0xE8, 0x0F), p(0xE8, 0x13), p(0xE8, 0x17), p(0xE8, 0x23), p(0xE8, 0x27), p(0xE8, 0x43), p(0xE9, 0x03),
-        p(0xEA, 0x03), p(0xEB, 0x01), p(0xEB, 0x40), p(0xEB, 0x61), p(0xEB, 0xA1), p(0xEB, 0xC1), p(0xEB, 0xE1), p(0xED, 0x7C),
-        p(0xF0, 0x08), p(0xF1, 0x7D), p(0xF3, 0x03), p(0xF3, 0x53), p(0xF4, 0x03), p(0xF5, 0x03), p(0xF5, 0x5B), p(0xF5, 0x7B),
-        p(0xF6, 0x03), p(0xF7, 0x03), p(0xF7, 0x63), p(0xF8, 0x03), p(0xF9, 0x00), p(0xF9, 0x01), p(0xF9, 0x02), p(0xF9, 0x03),
-        p(0xF9, 0x08), p(0xF9, 0x09), p(0xF9, 0x0A), p(0xF9, 0x1F), p(0xF9, 0x20), p(0xF9, 0x21), p(0xF9, 0x28), p(0xF9, 0x29),
-        p(0xF9, 0x2A), p(0xF9, 0x3F), p(0xF9, 0x40), p(0xF9, 0x48), p(0xF9, 0x68), p(0xF9, 0x69), p(0xF9, 0x6B), p(0xF9, 0x7F),
-        p(0xF9, 0x88), p(0xF9, 0xA8), p(0xF9, 0xC8), p(0xF9, 0xE0), p(0xF9, 0xE1), p(0xF9, 0xE2), p(0xF9, 0xE8), p(0xF9, 0xE9),
-        p(0xF9, 0xEA), p(0xF9, 0xF3), p(0xF9, 0xF4), p(0xF9, 0xF5), p(0xF9, 0xF7), p(0xF9, 0xFB), p(0xF9, 0xFF), p(0xFA, 0x03),
-        p(0xFA, 0xFF), p(0xFB, 0x73), p(0xFB, 0xFF), p(0xFC, 0xFF), p(0xFD, 0xFF), p(0xFE, 0x0B), p(0xFE, 0xFF), p(0xFF, 0x03),
-        p(0xFF, 0x17), p(0xFF, 0x43), p(0xFF, 0x54), p(0xFF, 0x83), p(0xFF, 0x97), p(0xFF, 0xB4), p(0xFF, 0xC3), p(0xFF, 0xFF),
-    });
-
-    static constexpr inline auto scores_x1 = std::to_array<std::uint16_t>({
-        0x002, 0x040, 0x0D9, 0x04F, 0x0A3, 0x00E, 0x06F, 0x12A, 0x048, 0x0F0, 0x07E, 0x0AE, 0x06E, 0x032, 0x05F, 0x001,
-        0x011, 0x0F4, 0x00D, 0x087, 0x006, 0x043, 0x092, 0x018, 0x086, 0x024, 0x0B8, 0x019, 0x08F, 0x0B5, 0x074, 0x01B,
-        0x004, 0x0ED, 0x00A, 0x078, 0x14C, 0x073, 0x021, 0x071, 0x0C9, 0x18B, 0x1C0, 0x08C, 0x01C, 0x1A0, 0x05E, 0x046,
-        0x10B, 0x027, 0x0A2, 0x05D, 0x114, 0x156, 0x112, 0x145, 0x110, 0x0A6, 0x016, 0x0EE, 0x12C, 0x137, 0x01E, 0x0CA,
-        0x04E, 0x0DF, 0x0C6, 0x1A6, 0x009, 0x028, 0x05C, 0x0CF, 0x149, 0x1E7, 0x057, 0x16E, 0x013, 0x01A, 0x02A, 0x033,
-        0x054, 0x076, 0x0A1, 0x0D3, 0x106, 0x171, 0x010, 0x072, 0x036, 0x0DB, 0x081, 0x126, 0x195, 0x050, 0x14E, 0x15D,
-        0x0B7, 0x184, 0x00C, 0x155, 0x03E, 0x0E2, 0x04C, 0x066, 0x1D5, 0x101, 0x04A, 0x056, 0x1FE, 0x152, 0x02E, 0x026,
-        0x058, 0x15A, 0x17B, 0x16D, 0x0D1, 0x0E6, 0x119, 0x1D7, 0x12D, 0x0CC, 0x13D, 0x150, 0x098, 0x136, 0x10E, 0x089,
-        0x038, 0x096, 0x191, 0x08B, 0x0EC, 0x138, 0x1FD, 0x03B, 0x077, 0x1FF, 0x06A, 0x190, 0x068, 0x09E, 0x12B, 0x0D2,
-        0x1A2, 0x133, 0x0E3, 0x0C5, 0x097, 0x117, 0x153, 0x0B2, 0x0F6, 0x0FB, 0x0E4, 0x0A9, 0x0D5, 0x014, 0x179, 0x01F,
-        0x120, 0x1EB, 0x11D, 0x130, 0x02C, 0x180, 0x1E3, 0x03D, 0x0D4, 0x1E2, 0x0DC, 0x1CF, 0x064, 0x05A, 0x16A, 0x093,
-        0x129, 0x0BF, 0x199, 0x13A, 0x0F9, 0x100, 0x108, 0x15F, 0x1CB, 0x11A, 0x03F, 0x015, 0x053, 0x1B9, 0x1EC, 0x02D,
-        0x0E1, 0x167, 0x1CA, 0x025, 0x045, 0x10F, 0x146, 0x080, 0x052, 0x11B, 0x065, 0x13C, 0x04D, 0x1CD, 0x0DD, 0x169,
-        0x139, 0x0AB, 0x10C, 0x0E0, 0x095, 0x0FA, 0x18F, 0x1DF, 0x148, 0x16B, 0x1F1, 0x19E, 0x18D, 0x1C3, 0x1F9, 0x0A4,
-        0x1E1, 0x14F, 0x1E4, 0x161, 0x061, 0x160, 0x085, 0x0A7, 0x1F5, 0x029, 0x0D0, 0x1DD, 0x143, 0x0B3, 0x1A1, 0x007,
-        0x1CE, 0x144, 0x000, 0x1B3, 0x079, 0x1FB, 0x07B, 0x0C7, 0x15B, 0x039, 0x0BA, 0x0C0, 0x1AC, 0x0AD, 0x088, 0x111,
-        0x1ED, 0x103, 0x1A8, 0x165, 0x1BD, 0x1AD, 0x16C, 0x1AF, 0x0D6, 0x1BE, 0x19D, 0x173, 0x1DC, 0x13B, 0x1C7, 0x09D,
-        0x0FF, 0x0B0, 0x172, 0x1A4, 0x187, 0x124, 0x17A, 0x044, 0x082, 0x125, 0x170, 0x1FA, 0x1C1, 0x0D7, 0x037, 0x185,
-        0x11F, 0x142, 0x0A5, 0x14A, 0x0AA, 0x192, 0x0AF, 0x183, 0x1DE, 0x198, 0x197, 0x19C, 0x182, 0x132, 0x109, 0x188,
-        0x1DA, 0x164, 0x069, 0x003, 0x0F1, 0x11C, 0x06B, 0x131, 0x1CC, 0x1B6, 0x0FD, 0x17D, 0x196, 0x1D3, 0x17E, 0x1EA,
-        0x116, 0x1F7, 0x0C1, 0x0CB, 0x1A7, 0x099, 0x020, 0x02F, 0x07F, 0x0EB, 0x1F8, 0x094, 0x1D1, 0x1BA, 0x1AA, 0x151,
-        0x060, 0x123, 0x063, 0x09A, 0x1B5, 0x09B, 0x186, 0x075, 0x1A3, 0x0B4, 0x1C6, 0x11E, 0x113, 0x166, 0x14D, 0x1D2,
-        0x0AC, 0x1DB, 0x1F0, 0x06D, 0x03A, 0x083, 0x0D8, 0x0B6, 0x070, 0x12E, 0x134, 0x1BB, 0x1B0, 0x059, 0x1BC, 0x07C,
-        0x030, 0x049, 0x0C4, 0x121, 0x1B1, 0x09C, 0x14B, 0x12F, 0x13F, 0x1E8, 0x1F3, 0x1E6, 0x15C, 0x1E0, 0x122, 0x105,
-        0x062, 0x154, 0x07D, 0x1B8, 0x1C9, 0x1C4, 0x0E8, 0x1F6, 0x0DE, 0x128, 0x174, 0x19B, 0x035, 0x08A, 0x15E, 0x1B7,
-        0x0BC, 0x19F, 0x1C5, 0x1D9, 0x0B1, 0x127, 0x0E5, 0x1D8, 0x005, 0x0F2, 0x194, 0x0C8, 0x0F7, 0x1D0, 0x0CD, 0x135,
-        0x00F, 0x118, 0x0F8, 0x193, 0x104, 0x163, 0x01D, 0x177, 0x042, 0x1C8, 0x09F, 0x0DA, 0x13E, 0x1E5, 0x034, 0x0FC,
-        0x1BF, 0x181, 0x176, 0x1C2, 0x168, 0x10A, 0x189, 0x0C3, 0x18C, 0x1AE, 0x102, 0x1D4, 0x18E, 0x1AB, 0x1FC, 0x140,
-        0x1EE, 0x0BB, 0x03C, 0x023, 0x04B, 0x067, 0x041, 0x162, 0x091, 0x0BE, 0x084, 0x0E9, 0x051, 0x0C2, 0x1F4, 0x107,
-        0x012, 0x06C, 0x157, 0x022, 0x18A, 0x1EF, 0x0A0, 0x0BD, 0x1E9, 0x10D, 0x1A9, 0x158, 0x0A8, 0x1F2, 0x0B9, 0x17C,
-        0x0FE, 0x16F, 0x1B2, 0x02B, 0x047, 0x115, 0x031, 0x090, 0x1B4, 0x08D, 0x175, 0x0EF, 0x17F, 0x19A, 0x07A, 0x1A5,
-        0x1D6, 0x147, 0x178, 0x0F3, 0x08E, 0x159, 0x055, 0x0CE, 0x008, 0x0F5, 0x017, 0x0E7, 0x05B, 0x141, 0x0EA, 0x00B,
-    });
-}
